% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_queryMass.R
\name{annotateMS1}
\alias{annotateMS1}
\title{xcms annotation of MS1}
\usage{
annotateMS1(x, mode = c("pos", "neg")[1], ref, ppmtol = 10,
  ips = 0.75, fun_parallel = parallel::mclapply, ...)
}
\arguments{
\item{x}{an object of prunedXcmsSet}

\item{mode}{Ionization mode. pos or neg}

\item{ref}{reference MS2 spectra used annotate the experimental MS2 spectra}

\item{ppmtol}{the mass tolerence given by parts per million (PPM)}

\item{ips}{The likelihood of the rule ("ips" column) in adduct table in MAIT package.}

\item{fun_parallel}{the parallel function, could be \code{mclapply} or \code{bplapply}}

\item{...}{the parameters passed to parallel function}
}
\description{
xcms annotation of MS1
}
