% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotateMS2.R
\name{annotateMS2}
\alias{annotateMS2}
\title{Annotate using MS2 spectra}
\usage{
annotateMS2(x, ms1Annot, ppmtol = 20)
}
\arguments{
\item{x}{a prunedXcmsSet object}

\item{ms1Annot}{the MS1 annotation table returned by \code{annotateMS1}}

\item{ppmtol}{the mass tolerence given by parts per million (PPM)}
}
\description{
Annotate using MS2 spectra
}
