% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chromPeaksPostFilter.R
\name{chromPeaksPostFilter}
\alias{chromPeaksPostFilter}
\title{post filter of chromPeaks}
\usage{
chromPeaksPostFilter(
  x,
  mtab,
  itab,
  postfilter = c(5, 500),
  BPPARAM = bpparam()
)
}
\arguments{
\item{x}{An XCMSnExp object after peak calling}

\item{mtab}{scan meta table returned by \link{asMetaTable}}

\item{itab}{intensity table returned by \link{asIntensityTable}}

\item{postfilter}{a numeric vector of length 2, in for mat c(x, y), meaning at least x peaks with intensity higher than y.}

\item{BPPARAM}{arguments passed to bplapply}
}
\description{
post filter of chromPeaks
}
