% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spectraComparison.R
\name{cosnorm}
\alias{cosnorm}
\title{normalized cosine distance between two positve vectors}
\usage{
cosnorm(a, b, n = 1)
}
\arguments{
\item{a}{a non-negative numerical vector}

\item{b}{the second non-negative numerical vector has the same length as a}

\item{n}{power factor for a and b, a and b will be transformed to a^n and b^n}
}
\value{
the normalized cosine distance
}
\description{
normalized cosine distance between two positve vectors
}
\details{
if a and b are non-negative, the expected cosine distance between a and b will
 not be 0. Assuming the elements in a and b are random, the normalized cosine
 distance is calculated as:
   observ <- sum(a*b)/(sqrt(sum(a^2))*sqrt(sum(b^2)))
   expect <- sum(mean(a)*mean(b)*length(a))/(sqrt(sum(a^2))*sqrt(sum(b^2)))
   (observ-expect)/(1-expect)
}
\examples{
 x <- abs(rnorm(50))
 y <- abs(rnorm(50))
 cosnorm(x, y)
}
