% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mirrorPlot.r
\name{.mirrorPlot}
\alias{.mirrorPlot}
\title{internal function used by mirrorPlotModule}
\usage{
.mirrorPlot(
  x,
  col,
  legend.peak.lower = "",
  legend.peak.upper = "",
  highlight = NULL,
  xlim = NULL
)
}
\arguments{
\item{x}{data.frame with columne mz and intensity}

\item{col}{peak colors}

\item{legend.peak.lower}{the legend for peak.lower intensities}

\item{legend.peak.upper}{the legend forpeak.upper intensities}

\item{highlight}{the peak to be peak.lower}

\item{xlim}{xlim passed to plot}
}
\description{
internal function used by mirrorPlotModule
}
