% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mirrorPlot.r
\name{.prep_mirrorPlot}
\alias{.prep_mirrorPlot}
\title{internal function used by mirrorPlotModule}
\usage{
.prep_mirrorPlot(peak.upper, peak.lower = NULL, ppmtol = 10)
}
\arguments{
\item{peak.upper}{the upper peak}

\item{peak.lower}{the lower peak}

\item{ppmtol}{mass tolerance (in ppm) comparing the upper and lower peaks}
}
\description{
internal function used by mirrorPlotModule
}
