% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_EIC.R
\name{EIC_module}
\alias{EIC_module}
\title{Module for EIC}
\usage{
EIC_module(input, output, session, pdata, object, obj,
  switchOnRepPeak = 12, file = reactive(NULL))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{pdata}{phenotype data}

\item{object}{xcmsViewer object}

\item{obj}{object returned by "xcmsAnnotationTab_module", the list object when a feature is selected,
it should contain at list an element named "eic_param", which gives the range of mz (numeric vector of 
length two; named mz) and range of RT (numeric vector of length two; named rt).}

\item{switchOnRepPeak}{Integer; if the numbe of sample is large, then only the representative
EICs are shown. This parameter tells how many samples will be considered as "large size" of samples.}

\item{file}{integer; file index}
}
\description{
Module for EIC
}
