% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_featurePostFilter.R
\name{featurePostfilter}
\alias{featurePostfilter}
\title{post filter of features}
\usage{
featurePostfilter(x, nComplete = 1)
}
\arguments{
\item{x}{An XCMSnExp object after peak calling}

\item{nComplete}{the number of complete samples in a 
group (is length = 1) /all groups (if length = group numbers). If can be given
integers (>= 1) or the percent of groups size (0, 1).#'}
}
\description{
post filter of features
}
