% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_asTableList.R
\name{filterSpectrum}
\alias{filterSpectrum}
\title{Filter MS spectrum}
\usage{
<<<<<<< HEAD
filterSpectrum(x, noise = 30, maxPeaks = 100, maxIdenticalInt = 5,
  maxMZ = Inf)
=======
filterSpectrum(x, noise = 30, maxPeaks = 100, maxIdenticalInt = 5, maxMZ = Inf)
>>>>>>> master
}
\arguments{
\item{x}{An Spectrum object}

\item{noise}{the noise level}

\item{maxPeaks}{maximum number of peaks to retain}

\item{maxIdenticalInt}{maximum number of idential intensities (often noise)}

\item{maxMZ}{maximum MZ, peaks at higher mz will be removed first, used when filter MS2 
spectra to limit mz lower than mass of parent ion}
}
\description{
Filtering noise in a spectrum.
}
