% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_zzz.R
\name{formatFeaturePeakTab}
\alias{formatFeaturePeakTab}
\title{format feature or peak table}
\usage{
formatFeaturePeakTab(tab)
}
\arguments{
\item{tab}{table}
}
\description{
format feature or peak table
}
