% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chromPeaksPostFilter.R
\name{getMetaIntensityTable}
\alias{getMetaIntensityTable}
\title{Extract scan meta and intensity table from XCMSnExp}
\usage{
getMetaIntensityTable(
  x,
  ms1.noise = 100,
  ms1.maxPeaks = Inf,
  ms1.maxIdenticalInt = 20,
  ms2.noise = 30,
  ms2.maxPeaks = 100,
  ms2.maxIdenticalInt = 6,
  ...
)
}
\arguments{
\item{x}{An XCMSnExp object after peak calling}

\item{ms1.noise}{passed to \link{asIntensityTable}}

\item{ms1.maxPeaks}{passed to \link{asIntensityTable}}

\item{ms1.maxIdenticalInt}{passed to \link{asIntensityTable}}

\item{ms2.noise}{passed to \link{asIntensityTable}}

\item{ms2.maxPeaks}{passed to \link{asIntensityTable}}

\item{ms2.maxIdenticalInt}{passed to \link{asIntensityTable}}

\item{...}{passed to \code{\link[MSnbase]{spectrapply}} or \code{\link[BiocParallel]{bplapply}}}
}
\value{
a list of "mtab" and "itab"
}
\description{
Extract scan meta and intensity table from XCMSnExp
}
