% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mirrorPlot.r
\name{mirrorPlot}
\alias{mirrorPlot}
\title{Mirror plot of MS2 fragmentation patterns}
\usage{
mirrorPlot(
  peak.upper,
  peak.lower,
  legend.peak.lower = "",
  legend.peak.upper = "",
  ppmtol = 10,
  highlight = NULL,
  xlim = NULL
)
}
\arguments{
\item{peak.upper}{the peak.upper fragmentation pattern, a data.frame with two columns named "mz" and "intensity"}

\item{peak.lower}{the peak.lower fragmentation pattern, a data.frame with two columns named "mz" and "intensity"}

\item{legend.peak.lower}{the legend for peak.lower intensities}

\item{legend.peak.upper}{the legend forpeak.upper intensities}

\item{ppmtol}{the mass tolerane used to map the intensity}

\item{highlight}{the peak to be peak.lower}

\item{xlim}{xlim passed to plot}
}
\description{
Mirror plot of MS2 fragmentation patterns
}
\examples{
df <- data.frame(
  mz = abs(rnorm(20)*200),
  intensity = abs(rnorm(20)*1000)
)
set.seed(1000)
me <- df[sample(1:nrow(df), size = 15), ] + rnorm(5, sd = 0.0005)
st <- df[sample(1:nrow(df), size = 10), ]

mirrorPlot(peak.lower = me,
           peak.upper = st,
           legend.peak.lower  = "xxxxxxxxxxx",
           legend.peak.upper = "yyyyyyyyyyyyy")

mirrorPlot(peak.lower = me,
           peak.upper = st,
           legend.peak.lower  = "xxxxxxxxxxx",
           legend.peak.upper = "yyyyyyyyyyyyy",
           xlim = c(50, 120))
}
