% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prepViewerData.R
\name{multi.t.test2}
\alias{multi.t.test2}
\title{Functions for multiple t-test}
\usage{
multi.t.test2(
  x,
  pheno,
  compare = NULL,
  median.center = FALSE,
  fillNA = TRUE,
  ...
)
}
\arguments{
\item{x}{the expression matrix, should be log10 transformed}

\item{pheno}{the phenotype data.frame}

\item{compare}{comparisons to be done. An nx3 matrix, the first colume indicate the 
phenotype column, the 2nd and 3rd columns indicate the two groups that should be compared}

\item{median.center}{logical; whether the expression matrix should be median centered}

\item{fillNA}{whether the NA should be filled}

\item{...}{other parameters passed to t.test}
}
\description{
Functions for multiple t-test
}
