% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prepViewerData.R
\name{pca}
\alias{pca}
\title{Function to perform PCA}
\usage{
pca(x, n = 6, prefix = "", fillNA = TRUE, samples = NULL)
}
\arguments{
\item{x}{the expression matrix}

\item{n}{number of components to include}

\item{prefix}{the prefix of headers in loading and PC matrix}

\item{fillNA}{logical}

\item{samples}{which samples to be included in the PCA}
}
\description{
the PCA result is ready to be included in the prunedXcmsSet
}
