% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPrunedXcmsSet.R
\name{peakPicking}
\alias{peakPicking}
\title{peak picking for individual pipeline}
\usage{
peakPicking(
  files,
  param,
  tmpdir = "xcmsViewerTemp",
  postfilterParam = list(postfilter = c(5, 1000), ms1.noise = 100, ms1.maxPeaks = Inf,
    ms1.maxIdenticalInt = 20, ms2.noise = 30, ms2.maxPeaks = 100, ms2.maxIdenticalInt =
    6, BPPARAM = bpparam())
)
}
\arguments{
\item{files}{a character vector stores the the mzXML or mzML files}

\item{param}{peak picking parameters, passed to \code{findChromPeaks}}

\item{tmpdir}{the temporary directory to store the processed results}

\item{postfilterParam}{passed to \code{chromPeaksPostFilter}}
}
\value{
it returns a list of three elements:
 1. itab = file paths of intensity tables
 2. mtab = file  paths of the meta tables
 3. XCMSnExp = file paths of XCMSnExp
 The elements in the returned list could be passed to \code{defineFeatures}
}
\description{
peaks picking are peroformed one raw file after another so that less memory
  used.
}
