% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plotlyEIC.R
\name{plotly_EIC}
\alias{plotly_EIC}
\title{Plot EIC using plotly}
\usage{
plotly_EIC(x, range, xlim = NULL, legend.title = "Group", col = NA)
}
\arguments{
\item{x}{a data.frame with 5 columns
- mz: mz
- intensity: intensity
- file: integer, the file index
- colorGroup: character
- tooltips: character tooltips}

\item{range}{a named numeric vector with three values named as "rtmin", "rtmax" and "rtmed"}

\item{xlim}{xlim}

\item{legend.title}{title of legend}

\item{col}{color code for EICs, it needs to be a named vector of colors, the names should be 
the unique values of "x$colorGroup".}
}
\description{
Plot EIC using plotly
}
