% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prepareAllMGFs.R
\name{prepareAllMGFs}
\alias{prepareAllMGFs}
\title{prepare MGF file to be download}
\usage{
prepareAllMGFs(obj, file)
}
\arguments{
\item{obj}{an object of prunedXcmsSet, returned by prepVeiwerData}

\item{file}{output file}
}
\description{
prepare MGF file to be download
}
