% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_zzz.R
\name{prepareMGF}
\alias{prepareMGF}
\title{prepare MGF file to be download}
\usage{
prepareMGF(scan, scanMeta, cons, mode, featureId)
}
\arguments{
\item{scan}{scan list, a data.frame with three columns - mz, intensity, ID}

\item{scanMeta}{scan meta data}

\item{cons}{consensusMS2Spectrum - a data.frame with two columns, mz and intensity}

\item{mode}{1+ or 1-}

\item{featureId}{ID}
}
\description{
prepare MGF file to be download
}
\examples{
# x is obj() from xcmsAnnotationTab_module
# mgf <- prepareMGF(
# scan = x$ms2scan,
# scanMeta = x$ms2scanMeta,
# cons = x$consensusMS2Spectrum,
# mode = "pos",
# featureId = x$featureid
# )
# writeLines(mgf, con = "spectrum.mfg")

}
