% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prepRef.R
\name{prepRef}
\alias{prepRef}
\title{preprare reference}
\usage{
prepRef(ref, mode, primaryInChI = NULL)
}
\arguments{
\item{ref}{annotation data.frame}

\item{mode}{mode, either pos or neg}

\item{primaryInChI}{the InChIkey for primary annotation, not used yet, so 
users need to edit the ref before passed to the runPrunedXcmsSet beforehand.}
}
\description{
preprare reference
}
