% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prepViewerData.R
\name{prepViewerData}
\alias{prepViewerData}
\title{Function to perform multiple PCA}
\usage{
prepViewerData(
  object,
  pheno = NULL,
  median.center = FALSE,
  fillNA = TRUE,
  compare.t.test = NULL,
  compare.pca = list(`_all_` = TRUE),
  nf = 6,
  ...,
  fx = "General|All|rtmed",
  fy = "General|All|mzmed",
  sx = "PCA|All|PC1 (",
  sy = "PCA|All|PC2 ("
)
}
\arguments{
\item{object}{the prunedXcmsSet object}

\item{pheno}{the phenotype data.frame, could be NULL, which means pData from object will be used. 
It is suggested to have a column named as "label", which contains unique value for sample labels.}

\item{median.center}{logical; whether to median center the data}

\item{fillNA}{logical; whether to fill the missing value}

\item{compare.t.test}{see \code{multi.t.test2}}

\item{compare.pca}{see \code{multi.pca}}

\item{nf}{number of components to include}

\item{...}{other parameters passed to t.test}

\item{fx}{the default x-axis of feature space, e.g "ttest|A_vs_B|mean.diff", for volcano plot}

\item{fy}{the default y-axis of feature space, e.g "ttest|A_vs_B|log.pvalue" or "ttest|A_vs_B|log.fdr", for volcano plot}

\item{sx}{the default x-axis of sample space}

\item{sy}{the default y-axis of sample space}
}
\description{
the PCA result is ready to be included in the prunedXcmsSet
}
