% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R, R/all_generics.R
\docType{class}
\name{prunedXcmsSet-class}
\alias{prunedXcmsSet-class}
\alias{dumpPrunedXcmsSet}
\alias{dumpPrunedXcmsSet,prunedXcmsSet-method}
\alias{annotateMetabolite}
\alias{annotateMetabolite,prunedXcmsSet-method}
\alias{exportTables}
\alias{exportTables,prunedXcmsSet-method}
\alias{exportTables,SQLiteConnection-method}
\title{prunedXcmsSet S4 class}
\usage{
dumpPrunedXcmsSet(object, ...)

\S4method{dumpPrunedXcmsSet}{prunedXcmsSet}(object, db.file, overwrite = FALSE)

annotateMetabolite(object, ...)

\S4method{annotateMetabolite}{prunedXcmsSet}(
  object,
  mode = c("pos", "neg")[1],
  ref,
  ppmtol = 25,
  ips = 0.75,
  fun_parallel = parallel::mclapply,
  ...
)

exportTables(object, file)

\S4method{exportTables}{prunedXcmsSet}(object, file)

\S4method{exportTables}{SQLiteConnection}(object, file)
}
\arguments{
\item{object}{an prunedXcmsSet object}

\item{...}{other arguments, such as}

\item{db.file}{the directory and name of the targe database file and}

\item{overwrite}{overwirte the target database if existing}

\item{mode}{Ionization mode. pos or neg}

\item{ref}{reference MS2 spectra used annotate the experimental MS2 spectra}

\item{ppmtol}{the mass tolerence given by parts per million (PPM)}

\item{ips}{ips}

\item{fun_parallel}{parallel function, often either mclapply or biocparallel}

\item{file}{xlsx file}
}
\description{
prunedXcmsSet
}
\section{Slots}{

\describe{
\item{\code{featureSet}}{xcmsFeatureSet}

\item{\code{peak}}{xcmsPeak}

\item{\code{scan}}{xcmsScan}

\item{\code{annot}}{annotation}

\item{\code{EIC}}{extracted ion currents}
}}

