% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_getX.R
\name{prunedXcmsSetAccesser}
\alias{prunedXcmsSetAccesser}
\alias{getMassTol}
\alias{getMassTol,prunedXcmsSet-method}
\alias{getMassTol,SQLiteConnection-method}
\alias{getScanIDFromFeatureID}
\alias{getFeatureMeta}
\alias{getFeatureMeta,prunedXcmsSet-method}
\alias{getFeatureMeta,SQLiteConnection-method}
\alias{getAnnot}
\alias{getAnnot,prunedXcmsSet,missing-method}
\alias{getAnnot,prunedXcmsSet,character-method}
\alias{getAnnot,SQLiteConnection,missing-method}
\alias{getAnnot,SQLiteConnection,character-method}
\alias{getPheno}
\alias{getPheno,prunedXcmsSet-method}
\alias{getPheno,SQLiteConnection-method}
\alias{getFeatureExprs}
\alias{getFeatureExprs,prunedXcmsSet-method}
\alias{getFeatureExprs,SQLiteConnection-method}
\alias{getPeak}
\alias{getPeak,prunedXcmsSet-method}
\alias{getPeak,SQLiteConnection-method}
\alias{getScan2}
\alias{getScan2,prunedXcmsSet-method}
\alias{getScan2,SQLiteConnection-method}
\alias{getScanMeta}
\alias{getScanMeta,prunedXcmsSet-method}
\alias{getScanMeta,SQLiteConnection-method}
\alias{getScanUniqueID}
\alias{getScanUniqueID,prunedXcmsSet-method}
\alias{getScanUniqueID,SQLiteConnection-method}
\alias{getAbundance}
\alias{getAbundance,prunedXcmsSet-method}
\alias{getAbundance,SQLiteConnection-method}
\alias{getEIC2}
\alias{getEIC2,list-method}
\alias{getEIC2,prunedXcmsSet-method}
\alias{getEIC2,SQLiteConnection-method}
\alias{getEICFromFeatureID}
\alias{getEICFromFeatureID,prunedXcmsSet-method}
\alias{getEICFromFeatureID,SQLiteConnection-method}
\alias{getMasterPeak}
\alias{getMasterPeak,xcmsFeatureSet-method}
\alias{getMasterPeak,prunedXcmsSet-method}
\alias{getIonMode}
\alias{getIonMode,SQLiteConnection-method}
\alias{getIonMode,prunedXcmsSet-method}
\title{Accessing elements in \code{\link[xcmsViewer]{prunedXcmsSet-class}}}
\usage{
getMassTol(object)

\S4method{getMassTol}{prunedXcmsSet}(object)

\S4method{getMassTol}{SQLiteConnection}(object)

getScanIDFromFeatureID(object, ID)

getFeatureMeta(object, ID = NULL)

\S4method{getFeatureMeta}{prunedXcmsSet}(object, ID = NULL)

\S4method{getFeatureMeta}{SQLiteConnection}(object, ID = NULL)

getAnnot(object, ID)

\S4method{getAnnot}{prunedXcmsSet,missing}(object, ID)

\S4method{getAnnot}{prunedXcmsSet,character}(object, ID)

\S4method{getAnnot}{SQLiteConnection,missing}(object, ID)

\S4method{getAnnot}{SQLiteConnection,character}(object, ID)

getPheno(object)

\S4method{getPheno}{prunedXcmsSet}(object)

\S4method{getPheno}{SQLiteConnection}(object)

getFeatureExprs(object, normalized = FALSE)

\S4method{getFeatureExprs}{prunedXcmsSet}(object, normalized = FALSE)

\S4method{getFeatureExprs}{SQLiteConnection}(object, normalized = FALSE)

getPeak(object, rowid = NULL)

\S4method{getPeak}{prunedXcmsSet}(object, rowid = NULL)

\S4method{getPeak}{SQLiteConnection}(object, rowid = NULL)

getScan2(object, scanId)

\S4method{getScan2}{prunedXcmsSet}(object, scanId)

\S4method{getScan2}{SQLiteConnection}(object, scanId)

getScanMeta(object, scanId)

\S4method{getScanMeta}{prunedXcmsSet}(object, scanId)

\S4method{getScanMeta}{SQLiteConnection}(object, scanId)

getScanUniqueID(object)

\S4method{getScanUniqueID}{prunedXcmsSet}(object)

\S4method{getScanUniqueID}{SQLiteConnection}(object)

getAbundance(object, featureId = NULL)

\S4method{getAbundance}{prunedXcmsSet}(object, featureId = NULL)

\S4method{getAbundance}{SQLiteConnection}(object, featureId = NULL)

getEIC2(object, rt = c(0, Inf), mz = c(0, Inf), file, na.rm = TRUE)

\S4method{getEIC2}{list}(object, rt = c(0, Inf), mz = c(0, Inf), file, na.rm = TRUE)

\S4method{getEIC2}{prunedXcmsSet}(object, rt = c(0, Inf), mz = c(0, Inf), file, na.rm = TRUE)

\S4method{getEIC2}{SQLiteConnection}(object, rt = c(0, Inf), mz = c(0, Inf), file, na.rm = TRUE)

getEICFromFeatureID(object, ID)

\S4method{getEICFromFeatureID}{prunedXcmsSet}(object, ID)

\S4method{getEICFromFeatureID}{SQLiteConnection}(object, ID)

getMasterPeak(object)

\S4method{getMasterPeak}{xcmsFeatureSet}(object)

\S4method{getMasterPeak}{prunedXcmsSet}(object)

getIonMode(object)

\S4method{getIonMode}{SQLiteConnection}(object)

\S4method{getIonMode}{prunedXcmsSet}(object)
}
\arguments{
\item{object}{An object of class \code{\link[xcmsViewer]{prunedXcmsSet-class}} or
sqlite3 database saved by  \code{\link[xcmsViewer]{dumpPrunedXcmsSet}} 
function.}

\item{ID}{Metabolite feature ID}

\item{normalized}{get the raw exprs or normalized exprs}

\item{rowid}{the row of the peak table, set this argument if only a 
subset rows of table should be returned.}

\item{scanId}{scan ID}

\item{featureId}{feature ID}

\item{rt}{range of retention time}

\item{mz}{range of mz}

\item{file}{which file to extract}

\item{na.rm}{logical, whether remove NA values}
}
\description{
These functions are used in the xcmsViewer shiny app. The input could be 
either a \code{prunedXcmsSet} object or sqlite3 database saved by 
\code{\link[xcmsViewer]{dumpPrunedXcmsSet}} function.
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

