% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPrunedXcmsSet.R
\name{runPrunedXcmsSet}
\alias{runPrunedXcmsSet}
\title{Process raw file to generate the prunedXcmsSet}
\usage{
runPrunedXcmsSet(
  files,
  peakPickingParam = MatchedFilterParam(fwhm = 7.5),
  tmpdir = "xcmsViewerTemp",
  postfilterParam = list(postfilter = c(5, 1000), ms1.noise = 100, ms1.maxPeaks = Inf,
    ms1.maxIdenticalInt = 20, ms2.noise = 30, ms2.maxPeaks = 100, ms2.maxIdenticalInt =
    6, BPPARAM = bpparam()),
  keepMS1 = TRUE,
  pheno = NULL,
  RTAdjustParam = PeakGroupsParam(minFraction = 0.25, extraPeaks = 1, smooth = "loess",
    span = 0.5, family = "gaussian", peakGroupsMatrix = matrix(nrow = 0, ncol = 0),
    subset = integer(), subsetAdjust = c("average", "previous")),
  peakGroupParam = PeakDensityParam(seq_len(length(files))),
  fillMissing = FALSE,
  mode = c("pos", "neg")[1],
  ref = NULL,
  ppmtol = 25,
  mclapplyParam = list(fun_parallel = mclapply, mc.cores = 1),
  bplapplyParam = MulticoreParam(workers = 1),
  peakPickingObj = NULL
)
}
\arguments{
\item{files}{mzXML/mzML files}

\item{peakPickingParam}{methods passed to \code{\link[xcms]{findChromPeaks}}}

\item{tmpdir}{directory for temporary files}

\item{postfilterParam}{post filter params to filter identified peaks, passed to \code{\link{chromPeaksPostFilter}}}

\item{keepMS1}{logical value. If the MS1 intensities should be kept.}

\item{pheno}{phenotype data. It should be a data.frame with at least one column named as "file" to 
list all the files passed to the function}

\item{RTAdjustParam}{methods passed to \code{\link[xcms]{adjustRtime}}}

\item{peakGroupParam}{methods passed to \code{\link[xcms]{groupChromPeaks}} for peak grouping}

\item{fillMissing}{logical; whether fill the missing values \code{\link[xcms]{fillChromPeaks}}}

\item{mode}{Ionization mode, either 'pos' (default) or "neg"}

\item{ref}{MS1 and MS2 annotation data}

\item{ppmtol}{the mass tolerence given by parts per million (PPM)}

\item{mclapplyParam}{the parallel function, could be \code{mclapply} or \code{bplapply}, used in peak picking and annotation}

\item{bplapplyParam}{the parameters passed to biocparallel biolapply, only use if fillmising is TRUE}

\item{peakPickingObj}{object save after the peak picking step, usually in "04_object_PeakPicking.RDS". If this is give the peak 
picking step will be ignored.}
}
\value{
An object of class "prunedXcmsSet"
}
\description{
Process raw file to generate the prunedXcmsSet
}
