% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_zzz.R
\name{str2spectra}
\alias{str2spectra}
\title{Convert spectra from string format to data.frame format}
\usage{
str2spectra(x)
}
\arguments{
\item{x}{a character (vector) of spectra instring format,
such as "78.454;123.333;332.232|50;1500;400"}
}
\value{
a data.frame with two columns "mz" and "intensity"
}
\description{
Convert spectra from string format to data.frame format
}
