% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{xcmsFeatureSet-class}
\alias{xcmsFeatureSet-class}
\alias{initialize,xcmsFeatureSet-method}
\title{xcmsFeatureSet S4 class that extends \code{"ExpressionSet"} class.}
\usage{
<<<<<<< HEAD
\S4method{initialize}{xcmsFeatureSet}(.Object, intensity, masterPeak,
  RTAdjustParam = NULL, peakGroupParam = NULL, ...)
=======
\S4method{initialize}{xcmsFeatureSet}(
  .Object,
  intensity,
  masterPeak,
  RTAdjustParam = NULL,
  peakGroupParam = NULL,
  ...
)
>>>>>>> master
}
\arguments{
\item{.Object}{object}

\item{intensity}{intensity}

\item{masterPeak}{maste peak}

\item{RTAdjustParam}{rt adjustment time}

\item{peakGroupParam}{peak group}

\item{...}{other parameter}
}
\description{
Some details
}
\section{Slots}{

\describe{
\item{\code{RTAdjustParam}}{Parameter used for retention time alignment}

\item{\code{peakGroupParam}}{Parameter used for peak grouping}
}}

