% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{xcmsPeak-class}
\alias{xcmsPeak-class}
\alias{initialize,xcmsPeak-method}
\title{xcmsPeak S4 class}
\usage{
\S4method{initialize}{xcmsPeak}(.Object, table, param, postFilter)
}
\arguments{
\item{.Object}{object}

\item{table}{table}

\item{param}{param}

\item{postFilter}{postFilter}
}
\description{
Peak tables extracted using xcms peakpicking
}
\section{Slots}{

\describe{
\item{\code{table}}{peak table}

\item{\code{param}}{Parameter used for peak picking}

\item{\code{postFilter}}{Parameter used for post filtering of peaks}
}}

