% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{xcmsScan-class}
\alias{xcmsScan-class}
\alias{initialize,xcmsScan-method}
\title{xcmsScan S4 class}
\usage{
\S4method{initialize}{xcmsScan}(.Object, meta, intensity, filter,
  keepMS1 = TRUE)
}
\arguments{
\item{.Object}{object}

\item{meta}{meta}

\item{intensity}{intensity}

\item{filter}{filter}

\item{keepMS1}{logical; whether MS1 scans should be kept}
}
\description{
scans
}
\section{Slots}{

\describe{
\item{\code{meta}}{scan meta}

\item{\code{intensity}}{The mz and intensities of scans}

\item{\code{filter}}{the filter used to filter the scans}

\item{\code{keepMS1}}{logical; whether MS1 scans are kept}
}}

