% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_xcmsViewer.R
\name{xcmsViewer}
\alias{xcmsViewer}
\title{Start xcmsViewer app}
\usage{
xcmsViewer(dir, ...)
}
\arguments{
\item{dir}{directory contain processed data as .RDS format, can be load into xcmsViewer}

\item{...}{current not used. In previous version, it was for other arguments passed to shiny::runApp}
}
\description{
Start xcmsViewer app
}
