% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chromPeaksPostFilter.R
\name{chromPeaksPostFilter}
\alias{chromPeaksPostFilter}
\title{post filter of chromPeaks}
\usage{
chromPeaksPostFilter(x, postfilter = c(5, 500), ms1.noise = 100,
  ms1.maxPeaks = Inf, ms1.maxIdenticalInt = 20, ms2.noise = 30,
  ms2.maxPeaks = 100, ms2.maxIdenticalInt = 6, ...)
}
\arguments{
\item{x}{An XCMSnExp object afater peak calling}

\item{postfilter}{= c(5, 500),}

\item{ms1.noise}{passed to \link{asIntensityTable}}

\item{ms1.maxPeaks}{passed to \link{asIntensityTable}}

\item{ms1.maxIdenticalInt}{passed to \link{asIntensityTable}}

\item{ms2.noise}{passed to \link{asIntensityTable}}

\item{ms2.maxPeaks}{passed to \link{asIntensityTable}}

\item{ms2.maxIdenticalInt}{passed to \link{asIntensityTable}}

\item{...}{arguments passed bplapply}
}
\description{
post filter of chromPeaks
}
